+++ packages/solana/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @turnkey/solana
 
+## 1.1.0-beta.0
+
+### Minor Changes
+
+- SDK beta release @turnkey/react-wallet-kit @turnkey/core
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-browser@5.9.0-beta.0
+  - @turnkey/sdk-server@4.7.0-beta.2
+  - @turnkey/core@1.0.0-beta.5
+  - @turnkey/http@3.10.0-beta.2
+
 ## 1.0.42
 
 ### Patch Changes
@@ -8,6 +22,51 @@
   - @turnkey/sdk-browser@5.8.0
   - @turnkey/sdk-server@4.7.0
   - @turnkey/http@3.10.0
+
+## 1.0.41-beta.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-server@4.7.0-beta.1
+  - @turnkey/http@3.10.0-beta.1
+  - @turnkey/sdk-browser@5.7.1-beta.1
+
+## 1.0.41-beta.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-server@4.7.0-beta.0
+  - @turnkey/http@3.10.0-beta.0
+  - @turnkey/sdk-browser@5.7.1-beta.0
+
+## 1.0.41-beta.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-browser@5.6.1-beta.2
+  - @turnkey/http@3.8.1-beta.2
+  - @turnkey/sdk-server@4.5.1-beta.2
+
+## 1.0.41-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-browser@5.6.1-beta.1
+  - @turnkey/http@3.8.1-beta.1
+  - @turnkey/sdk-server@4.5.1-beta.1
+
+## 1.0.41-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-browser@5.6.1-beta.0
+  - @turnkey/http@3.8.1-beta.0
+  - @turnkey/sdk-server@4.5.1-beta.0
 
 ## 1.0.41