+++ packages/viem/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @turnkey/viem
 
+## 0.14.0-beta.0
+
+### Minor Changes
+
+- SDK beta release @turnkey/react-wallet-kit @turnkey/core
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/api-key-stamper@0.5.0-beta.5
+  - @turnkey/sdk-browser@5.9.0-beta.0
+  - @turnkey/sdk-server@4.7.0-beta.2
+  - @turnkey/core@1.0.0-beta.5
+  - @turnkey/http@3.10.0-beta.2
+
 ## 0.13.0
 
 ### Minor Changes
@@ -19,6 +34,26 @@
 
 - [#848](https://github.com/tkhq/sdk/pull/848) [`8305dd2`](https://github.com/tkhq/sdk/commit/8305dd2cbfbb3cfc55ed37c9626198627c752fd9) Author [@andrewkmin](https://github.com/andrewkmin) - Fix: update serialization of typed data to conform to Viem's implementation
 
+## 0.12.1-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-server@4.7.0-beta.1
+  - @turnkey/http@3.10.0-beta.1
+  - @turnkey/sdk-browser@5.7.1-beta.1
+  - @turnkey/api-key-stamper@0.4.8-beta.4
+
+## 0.12.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-server@4.7.0-beta.0
+  - @turnkey/http@3.10.0-beta.0
+  - @turnkey/sdk-browser@5.7.1-beta.0
+  - @turnkey/api-key-stamper@0.4.8-beta.3
+
 ## 0.12.0
 
 ### Minor Changes
@@ -33,6 +68,36 @@
   - @turnkey/sdk-browser@5.7.0
   - @turnkey/sdk-server@4.6.0
   - @turnkey/http@3.9.0
+
+## 0.11.2-beta.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-browser@5.6.1-beta.2
+  - @turnkey/api-key-stamper@0.4.8-beta.2
+  - @turnkey/http@3.8.1-beta.2
+  - @turnkey/sdk-server@4.5.1-beta.2
+
+## 0.11.2-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-browser@5.6.1-beta.1
+  - @turnkey/api-key-stamper@0.4.8-beta.1
+  - @turnkey/http@3.8.1-beta.1
+  - @turnkey/sdk-server@4.5.1-beta.1
+
+## 0.11.2-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @turnkey/sdk-browser@5.6.1-beta.0
+  - @turnkey/api-key-stamper@0.4.8-beta.0
+  - @turnkey/http@3.8.1-beta.0
+  - @turnkey/sdk-server@4.5.1-beta.0
 
 ## 0.11.1